requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
dj-database-url==0.5.0
Django==3.1.5
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-phone-field==1.8.1
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
psycopg2==2.8.6
python-decouple==3.4
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
